Update omnibus configuration to point at ../ for oc_erchef
This is still uisng the git provider.  We could move to the path
provider to make it easier to build off an unclean working directory,
but this has proven significantly slower than a git-based build in
local tests.

The opscode-ci configuration will need to be cd into the omnibus
directory.